+++ packages/core/src/services/email/get-multiple.ts
@@ -38,6 +38,7 @@ const getMultiple: ServiceFn<
 			"last_error_message",
 			"last_attempt_at",
 			"last_success_at",
+			"simulate",
 			"created_at",
 		],
 		queryParams: data.query,