+++ packages/plugin-nodemailer/src/types/types.ts
@@ -5,5 +5,8 @@ export interface PluginOptions {
 		email: string;
 		name: string;
 	};
+	/** Your Nodemailer transporter instance */
 	transporter: Transporter;
+	/** When set to true, the plugin will not send emails but will still return as a success */
+	simulate?: boolean;
 }