+++ packages/plugin-nodemailer/CHANGELOG.md
@@ -1,11 +1,14 @@
 # @lucidcms/plugin-nodemailer
 
+## 1.1.2 (UNRELEASED)
+
+- Added new `simulate` config option so you can stop emails been sent out in a dev environment but have them still registered as a success.
+
 ## 1.1.1
 
 ### Patch Changes
 
 - Bumped core peer dependency version to `0.12.0-alpha.1`.
-
 
 ## 1.1.0