+++ packages/plugin-resend/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @lucidcms/plugin-resend
 
+## 1.0.2 (UNRELEASED)
+
+- Added new `simulate` config option so you can stop emails been sent out in a dev environment but have them still registered as a success.
+
 ## 1.0.1
 
 ### Patch Changes